requirements.txt
cycler==0.10.0
imutils==0.5.2
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
pdf2image==1.5.4
Pillow==6.0.0
pyparsing==2.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
six==1.12.0
utils==0.9.0
XlsxWriter==1.1.8
